cdc/requirements.txt
numpy
pandas
SQLAlchemy
python-dotenv
psycopg2-binary
